pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vdb-comparison"
version = "0.1.0"
description = "Vector database comparison benchmark"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pandas>=2.0.3",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "psycopg2-binary>=2.9.7",
    "qdrant-client>=1.7.0",
    "weaviate-client>=3.26.7,<4.0.0",
    "psutil>=5.9.5",
    "tqdm>=4.65.0",
    "scikit-learn>=1.3.0",
    "datasets>=2.0.0",
]

[project.scripts]
benchmark = "benchmark:main"
visualize = "visualize_results:main"
